feat: Add MAVLink communication via serial port
This commit adds the necessary code changes to establish MAVLink communication via the serial port. It includes the inclusion of the `HardwareSerial.h` and `MAVLink.h` libraries, as well as the initialization of the serial port with the appropriate settings. This enables communication with the Flight Controller (FC) using MAVLink.